Check name of global parameters before their values in GlobalParamMixin.__setattr__
When I tried to add a pandas.DataFrame as attribute to a task mixed with the
GlobalParamMixin, I encountered an exception because the DataFrame could not
be compared to the _no_default_value object, which makes sense. Checking the
name before the value ensure that this check is only performed for existing
global attributs.

Change-Id: Ie9762d5a53d3c23230101459d76802c44d444435